Moved substantially towards fully general build interception.
Appended instructions to the end of Readme for how to use the new
process; the old instructions are still there.

Added a makefile which uses 'which' and 'gcc' to compute which tools
to intercept.  Also makes softlinks for the multiple identities of the
shared interception scripts.

Added a makefile Intercept.mk which provides targets for easily
turning interception on and off.  Intercepted executables are now
simply postfixed with '_orig'.

Now have a gcc interceptor that
    1 - turns off combining the preprocessor with the compiler proper
    2 - points gcc at our spec file.
Note that we do not use the -B option, but instead override all the
tools directly.

Made cc1_interceptor more robust with respect to the -o option: you
can now have the space or not before the argument.  Same trick done in
the new collect2_interceptor.

Removed the comment about turning off -O in the build process.  Will
talk about it in the Readme.

We now record the arguments with which the scripts were called
delimited by newlines so they can be carefully replayed.

Added pass-through interceptors for cpp and as.


git-svn-id: http://build-interceptor.tigris.org/svn/build-interceptor/trunk@18 8700b9f2-3af6-0310-88e9-a4e0a7708a09